Refactor the evaluate function
The "evaluate" function was getting different outputs comparing with the
google online calculator due to a logic error. It was looping throught
the first matches, updating the output string and going to the next
operator, causing the string "123 + 123 + 123" to return "246 + 123"
instead of "369".
Also some regular expressions was refactored to remove redudancy, like
replacing /[\d]/ for only /\d/.